backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
sqlalchemy
psycopg2-binary
python-dotenv
alembic
httpx
bytez
python-jose[cryptography]
passlib[bcrypt]
python-multipart
